requirements.txt
fastapi
uvicorn
python-multipart
Pillow
tensorflow
supabase
python-dotenv